+++ Simple Game Library/META-INF/CHANGES.md
@@ -8,6 +8,7 @@ Date: *2016-09-??*
 - Improved vector classes, added Vector4d class.
 - Additional functionality for mouse input.
 - Add `setTitle` method to `Graphics` interface.
+- Added option to force graphics mode in `GfxConfigurator'.
 
 ## Version 1.4.1
 Date: *2016-06-06*